Add files via upload
Starting the Game
The final task for this part is to actually click the Start Game button and run the second scene in the game itself.

Adding Scenes to Build
Before you can run different scenes, you need to add them to the Scenes in Build list in the Project Settings, so that they are included in the final application.

To do this, on the menu select File \ Build Settings. This will open the Build Settings dialog. Then open the Scenes folder in the Project Browser and drag the MenuScene first, and then the RocketMouse scene to the Scenes in Build list.

53

Finally, close the Build Settings dialog.

Creating UIManager
When you add an event handler to the button, you need to specify which method to call when you click the button. Since you cannot use static methods, you will need to select a public method from a script that is attached to a GameObject.

From the top bar, choose GameObject \ Create Empty. Then select GameObject in the Hierarchy view and rename it to UIManager.

After that, click Add Component in the Inspector and select New Script. Name it UIManagerScript. Make sure the Language is set to CSharp and click Create and Add.

This is what you should see in the Hierarchy view and the Inspector view:

54

Double-click on the UIManagerScript in the Inspector to open the script in MonoDevelop. Once the script loads, remove both Start() and Update().

Next add the following statement underneath the previous `using` statements.

using UnityEngine.SceneManagement;

This allows you to load other scenes. Next add the following:

public void StartGame() 
{
    SceneManager.LoadScene("RocketMouse");
}

Save the script and move on to the next step:

Calling the StartGame method when the Player Clicks the Button
Switch back to Unity and follow these steps:

Select Btn_Start in the Hierarchy and scroll down in the Inspector to the On Click() list.
Click the + button to add new item.
Then drag UIManager from the Hierarchy to the newly added item in the list.
Click on the dropdown to select the function. Right now, it’s set to No Function.
In the opened menu select UIManagerScript and the select StartGame () in the menu that opens next.
55-new

Save Scenes your work and then Run the scene and click the Start Game button, this should open the game scene.